+++ test/src/Ouzo/Core/Db/QueryTest.php
@@ -105,4 +105,37 @@ class QueryTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(QueryType::$SELECT, $query->type);
         $this->assertEquals(array('name', 'id'), $query->selectColumns);
     }
+
+    /**
+     * @test
+     */
+    public function shouldLockForUpdate()
+    {
+        // when
+        $query = Query::select()->lockForUpdate();
+
+        // then
+        $this->assertTrue($query->lockForUpdate);
+    }
+
+    /**
+     * @test
+     */
+    public function selectShouldNotLockForUpdateByDefault()
+    {
+        // when
+        $query = Query::select();
+
+        // then
+        $this->assertFalse($query->lockForUpdate);
+    }
+
+    /**
+     * @expectedException \InvalidArgumentException
+     * @test
+     */
+    public function shouldNotAllowToUseLockForUpdateInNonSelectQueries()
+    {
+        Query::delete()->lockForUpdate();
+    }
 }