+++ test/src/Ouzo/Core/ModelTest.php
@@ -845,4 +845,20 @@ class ModelTest extends DbTransactionalTestCase
         //then
         $this->assertEquals('new name', $model->reload()->name);
     }
+
+    /**
+     * @test
+     */
+    public function shouldSelectForUpdate()
+    {
+        // given
+        Product::create(array('name' => 'Sport', 'price' => '123'));
+
+        // when
+        $products = Product::where()->lockForUpdate()->fetchAll();
+
+        // then
+        $this->assertCount(1, $products);
+    }
+
 }