+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Enhancements:
 * [Core] Added support for CSRF token in forms.
 * [Tools] Added method that lists all generated URI helper methods (GeneratedUriHelper::allGeneratedUriNames). 
 * [Utilities] Implemented Optional class (issue #72).
+* [ORM] Added support for [SELECT ... FOR UPDATE - with ModelQueryBuilder's lockForUpdate method](http://ouzo.readthedocs.org/en/latest/documentation/orm.html#locking).
 
 Bug fixes:
 * [Localization] Fixed I18n::loadLabels not to load translation file if it was already loaded (issue #173).